Restructure code + migration script + fix playActivityDB segfault
